# Make backtrace generation work outward from current frame