… vm_backtrace.c